pyproject.toml
[project]
name = "stonks-scraper"
version = "0.1.0"
description = "Automated pipeline for scraping, extracting, and aggregating financial statements from European company annual reports"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "openai>=1.12.0",
    "pdfplumber>=0.10.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
]
